backend/requirements.txt
fastapi>=0.115.5
uvicorn>=0.32.1
python-dotenv>=1.0.1
onnxruntime>=1.17.0
pinecone>=5.0.1
pydantic>=2.10.3
transformers>=5.0.0
psutil>=6.1.1
numpy>=2.0.0
pdfminer.six>=20231228
python-docx>=1.1.0
python-multipart>=0.0.9
